[GTK] Support HiDPI Properly in WebKitGtk+ with the TextureMapper
This patch fixes HiDPI issue in the TextureMapper.
To support HiDPI in the TextureMapper, we need to draw scaled contents
in the TextureMapperTile, and apply the global scale in the root layer
to apply transforms currectly.

Supporting the device scale is handled at TextureMapperLayer and
TextureMapperBackingStore, and GraphicsLayerTextureMapper do not handles
the device scale directly.

From the TextureMapperLayer, deviceScale and pageScale do not have to be
handled differently. These are multiplied and provided to
TextureMapperBackingStore.

Note: This patch is still working in progress.
There is a really ugly code in DrawingAreaImpl.cpp:86
I'm working on to avoid this issue.