Playing a video with other composited elements causes flickering.
GstVideoFrame was reused or invalidated while the compositing thread uses it
for compositing.  MediaPlayerPrivateGStreamerBase::triggerRepaint passed a
frame to the compositing thread after unmapping the previous frame.

The compositing thread can render a new scene without new frame from Gst if
there are other composited elements which invalidated. (e.g. CSS animation or
WebGL) It causes the decoding thread of Gst to recycle the frame which is still
used in the compositing thread.

    TextureMapperPlatformLayerBuffer::setUnmanagedBufferDestructor
Added to set a custom destructor function to the PlatformLayerBuffer. We can
unmap the video frame when the buffer is destucted from the compositing thread.
Because of that, we don't have to keep the video frame as a member of this
class.  Also, we can ensure the frame should be unmapped when it is not used at
compositing thread anymore.

This patch also modified MediaPlayerPrivateGStreamerBase to send video frame
from the decoding thread directly.  gst_video_frame_[un]map is a thread-safe.
so we can map a frame in the decoding thread, and unmap that frame in the
compositing thread without problem.

Note: Still needs some polishing to support the case without gst-gl.