Fixed a deadlock condition in GStreamerGL + Threaded Compositor
MediaPlayerPrivateGStreamerBase::updateOnCompositorThread and
TextureMapperPlatformLayerProxy::swapBuffer makes a dead lock condition.

TextureMapperPlatformLayerProxy::swapBuffer destroys previously used unmanaged
buffer in its lock. And the unmapping of this buffer (GstVideoFrame) also needs
a synchronization in the GStreamer's gl thread.

updateOnCompositorThread also needs to acquire a lock from
TextureMapperPlatformLayerProxy, and it already acquired the GStreamer's gl
thread.

Because the destruction of the previsouly used unmanagedbuffer don't have to be
happened in side of TextureMapperPlatformLayerProxy's lock, this patch defers
it to the out side of the lock to prevent the deadlock condition.