[ThreadedCompositor] Fix a crash happening when leaving a page including video.
When GstGL is disabled but threaded compositor is enabled, video has its own layer.
This layer is an independent texture from compositor's gl context. The problem is
the texture including the layer's gl context is destroyed in main thread
when leaving a page, while rendering thread does not know that and it keeps trying
swapBuffers with already destroyed texture.
To fix this crash, we defer destroying a layer of which texture is created
by its own gl context, not threaded compositor's gl context. we call such a texture 'managedTexture'.