[ThreadedCompositor] Post a video texture to the threaded compositor when entering fullscreen.
When entering fullscreen, webkit relayout and rebuild a render tree. However,
the media player does not have a change to update its texture and send it to the
compositor even though the video size to display is actually changed by the relayout.
This leads the threaded compositor to do nothing to update for video while compositing.
We need to send video textures to the compositor when the size of video changes.