[GTK] scroll with transparent background not repainted after scrollY >= 32768 https://bugs.webkit.org/show_bug.cgi?id=154283
Reviewed by NOBODY (OOPS!).

If the Cairo's matrix exceeds the limit of Pixman's maximum int size (which is
15bit), Cairo does not render bitmap because the given matrix is invalid.  When
we scroll a page which has a patterned bitmap as a background to more than the
Pixman's maximum int size, it simply doesn't render anything as background and
the target surface ends up with a transparent background.

This patch translates the CTM of Cairo's context when we drawing patterns to
the context to remove invisible area. Because it is a pattern there is no
visual difference between drawing tiles from the visible area and the beginning
of the page. But it is more natural to draw tiles from the visible area for the
performance.

* platform/graphics/cairo/CairoUtilities.cpp:
(WebCore::drawPatternToCairoContext):